Rework 3

+++ src/main/java/com/javamentor/qa/platform/dao/abstracts/model/VoteQuestionDao.java
@@ -3,8 +3,10 @@ package com.javamentor.qa.platform.dao.abstracts.model;
 import com.javamentor.qa.platform.models.entity.question.VoteQuestion;
 import com.javamentor.qa.platform.models.entity.user.User;
 
+import java.util.Optional;
+
 public interface VoteQuestionDao extends ReadWriteDao<VoteQuestion, Long>{
 
     boolean isUserVoteByQuestionIdAndUserId(Long id, Long userId);
-    int getVote(Long id);
+    Long getVote(Long id);
 }
